Use repository for archive creation
