+++ src/lib/dbSdk/index.ts
@@ -4,5 +4,4 @@ export * from './getArchivistBoundWitnessesMongoSdk'
 export * from './getArchivistPayloadMongoSdk'
 export * from './getBaseMongoSdk'
 export * from './getMongoDBValues'
-export * from './storeArchive'
 export * from './UpsertResult'
