+++ src/model/Repository/Repository.ts
@@ -1,15 +1,15 @@
 export interface QueryableRepository<TResponse, TQuery> {
   find(query: TQuery): Promise<TResponse>
 }
-export interface ReadRepository<TResponse, TId = string> {
-  get(id: TId): Promise<TResponse>
+export interface ReadRepository<TReadResponse, TId = string> {
+  get(id: TId): Promise<TReadResponse>
 }
 
-export interface WriteRepository<TResponse, TInsert> {
-  insert(item: TInsert): Promise<TResponse>
+export interface WriteRepository<TWriteResponse, TWrite> {
+  insert(item: TWrite): Promise<TWriteResponse>
 }
 
-export type ReadWriteRepository<TInsert, TResponse, TId = string> = ReadRepository<TResponse, TId> & WriteRepository<TResponse, TInsert>
+export type ReadWriteRepository<TWriteResponse, TWrite, TReadResponse = TWriteResponse, TId = string> = ReadRepository<TReadResponse, TId> & WriteRepository<TWriteResponse, TWrite>
 
 export type Repository<TInsert, TResponse, TId = string, TQuery = unknown> = ReadRepository<TResponse, TId> &
   WriteRepository<TResponse, TInsert> &
